[#34] FIX: 코드리뷰 반영
