+++ Hous-iOS-release/Global/UIComponent/GrayLineView.swift
@@ -9,13 +9,4 @@ import UIKit
 
 class GrayLineView: UIView {
   
-//  override func draw(_ rect: CGRect) {
-//    let path = UIBezierPath()
-//    UIColor(red: 245/255, green: 245/255, blue: 245/255, alpha: 1).set()
-//    path.move(to: CGPoint(x: 0, y: bounds.size.height))
-//    path.addLine(to: CGPoint(x: bounds.size.width, y: bounds.size.height))
-//    path.close()
-//    path.lineWidth = 3
-//    path.stroke()
-//  }
 }
